DSFAAP-749: add basic /task-list spike test
**Why task list**

The task-list page is one of the more computationally expensive ones -
since it calculates the validity of the entire application.

It also doesn't need any set up to be accessed, making it more amendable
to this testing.

Finally, as we iterate the service, this is likely to be the page that
changes the least as opposed to specific questions or specific sections,
which are likely to alter over time, keeping our perf tests stable

**Why the chosen numbers**

Private beta is for a max audience of 100 users.  The number of pages in
a journey is far south of 100 (even once we've added all the ones we
want to by the end of private beta, as far as we know right now).

If we simulate all of our participants completing journeys as fast as
(in)humanly possible - one HTTP request after another - and we suffer no
errors, then we know that we're going to be able to handle the load with
the scaling we currently have.